La lista no armaba: un tipo de dato mio, de hace una hora
Sergio creo "No han instalado la app" (con pedidos, menos los registrados),
toco Armar y enviar, y no salio nada. Se quedo en 0 contactos.

La causa es del cambio de las exclusiones, y es de las que no perdonan:

    fn_wa_criterio(..., p_n_pedidos integer, ...)
    v_wa_contactos.n_pedidos  →  numeric

Postgres NO convierte numeric a integer por su cuenta al elegir que funcion
llamar. Asi que no encontraba ninguna y `fn_wa_armar_lista` moria entera con
"function does not exist" — pero desde la pantalla eso se ve como una lista
vacia, o sea como "no hay a quien mandarle". Dos cosas muy distintas con la
misma cara, otra vez.

Y no era solo la lista nueva: cualquier lista habria fallado al rearmarse,
incluidas las que ya funcionaban. Por suerte no se mando nada mal — se mando
NADA.

Los tres numeros pasan a numeric: integer y bigint si entran solos, al reves
no. Comprobado sobre la lista de Sergio: 84 con pedidos, 54 quitando a los
registrados, y quedan armados los 54.

Lo que me llevo de aqui: al cambiar una funcion de base hay que llamarla contra
los datos REALES antes de decir que esta lista. La probe con valores escritos a
mano —donde el numero literal si se convierte solo— y por eso paso.

+++ sql/2026-08-28-listas-exclusiones.sql
@@ -27,7 +27,13 @@ create or replace function public.fn_wa_criterio(
   p_criterio      text,
   p_ya_escribio   boolean,
   p_guardado      boolean,
-  p_n_pedidos     integer,
+  --  NUMERIC y no INTEGER, y esto ya costo una lista que no salio.
+  --  `v_wa_contactos.n_pedidos` es numeric; Postgres NO convierte numeric a
+  --  integer por su cuenta al buscar la funcion, asi que la llamada no
+  --  encontraba ninguna y `fn_wa_armar_lista` moria con "function does not
+  --  exist". La lista quedaba en cero y parecia que no habia a quien mandarle.
+  --  Se declaran numeric los tres numeros: integer y bigint SI entran solos.
+  p_n_pedidos     numeric,
   p_tiene_nombre  boolean,
   p_registrado    boolean,
   p_instalada     boolean,
